+++ index.ts
@@ -1,15 +1,11 @@
 import 'dotenv/config';
 import { createServer, startServer, vendors, routers, port } from './server';
-import { notFoundHandler, errorHandler } from './server/middlewares';
+import { errorHandlers } from './server/middlewares';
 
 startServer(
   createServer({
     middlewares: [...vendors /*, middlewares */],
-    routes: [
-      ...routers,
-      { uri: '', handler: notFoundHandler, errorHandler: true },
-      { uri: '', handler: errorHandler, errorHandler: true },
-    ],
+    routes: [...routers, ...errorHandlers],
   }),
   port
 );