+++ server/index.ts
@@ -1,6 +1,6 @@
 import express, { Application } from 'express';
+import { Server } from 'http';
 import { ServerConfiguration } from '../types';
-import { requestLogger } from './middlewares';
 import { logger } from './vendors';
 
 export { port } from './constants';
@@ -13,12 +13,16 @@ export const createServer = (
   const server: Application = express();
 
   if (configuration) {
-    configuration.middlewares.forEach(middleware => server.use(middleware));
-    configuration.routes.forEach(route =>
-      route.errorHandler
-        ? server.use(route.handler, requestLogger)
-        : server.use(route.uri, route.handler, requestLogger)
-    );
+    if (configuration.middlewares.length > 0) {
+      configuration.middlewares.forEach(middleware => server.use(middleware));
+    }
+    if (configuration.routes.length > 0) {
+      configuration.routes.forEach(route =>
+        route.errorHandler
+          ? server.use(...route.handlers)
+          : server.use(route.uri, ...route.handlers)
+      );
+    }
   }
 
   return server;
@@ -27,8 +31,8 @@ export const createServer = (
 export const startServer = (
   server: Application,
   port: string | number
-): void => {
-  server.listen(port, () => {
+): Server => {
+  return server.listen(port, () => {
     logger.info('Server is listening on port ' + port);
   });
 };