+++ server/middlewares/index.ts
@@ -1,2 +1,6 @@
-export { notFoundHandler, errorHandler } from './errors';
+export {
+  default as errorHandlers,
+  notFoundHandler,
+  errorHandler,
+} from './errors';
 export { requestLogger, setLogging } from './logger';