+++ server/routes/index.ts
@@ -1,9 +1,10 @@
+import { requestLogger } from '../middlewares';
 import testRouter from './testEndpoint';
 
 export default [
   {
     uri: '/test',
-    handler: testRouter,
+    handlers: [testRouter, requestLogger],
     errorHandler: false,
   },
 ];